requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
sentence-transformers
transformers
faiss-cpu
joblib
pytest
httpx